% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nasdaq_data_link.R
\name{metaData}
\alias{metaData}
\title{Retrieve metadata from a Nasdaq Data Link series or search results}
\usage{
metaData(x)
}
\arguments{
\item{x}{A Nasdaq Data Link time series object or search results with attached meta data.}
}
\value{
Returns a list of meta data about the series or search results.
}
\description{
Retrieve metadata from a Nasdaq Data Link series or search results
}
\examples{
\dontrun{
metaData(ts)
}
}
\seealso{
\code{\link{NasdaqDataLink.api_key}}, \code{\link{NasdaqDataLink.search}}
}
