% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nasdaq_data_link_config.R
\name{NasdaqDataLink.api_key}
\alias{NasdaqDataLink.api_key}
\title{Query or set Nasdaq Data Link API key}
\usage{
NasdaqDataLink.api_key(api_key)
}
\arguments{
\item{api_key}{Optionally passed parameter to set Nasdaq Data Link \code{api_key}.}
}
\value{
Returns invisibly the currently set \code{api_key}.
}
\description{
Query or set Nasdaq Data Link API key
}
\examples{
\dontrun{
NasdaqDataLink.api_key('foobar')
}
}
