% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nasdaq_data_link_config.R
\name{NasdaqDataLink.auth}
\alias{NasdaqDataLink.auth}
\title{Query or set Nasdaq Data Link API token}
\usage{
NasdaqDataLink.auth(auth_token)
}
\arguments{
\item{auth_token}{Optionally passed parameter to set Nasdaq Data Link \code{auth_token}.}
}
\value{
Returns invisibly the currently set \code{auth_token}.
}
\description{
Deprecated. Alias of \code{\link{NasdaqDataLink.api_key}}
}
\examples{
\dontrun{
NasdaqDataLink.auth('foobar')
}
}
\seealso{
\code{\link{NasdaqDataLink.api_key}}
}
