% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nasdaq_data_link_config.R
\name{NasdaqDataLink.base_url}
\alias{NasdaqDataLink.base_url}
\title{Query or set Base URL for White-labeled sites}
\usage{
NasdaqDataLink.base_url(base_url)
}
\arguments{
\item{base_url}{Optionally passed parameter to set Nasdaq Data Link \code{base_url}.}
}
\value{
Returns invisibly the currently set \code{base_url}.
}
\description{
Query or set Base URL for White-labeled sites
}
\examples{
\dontrun{
NasdaqDataLink.base_url('http://localhost')
}
}
